+++ devices/temp_darwin.go
@@ -6,6 +6,7 @@ import smc "github.com/xxxserxxx/iSMC"
 
 func init() {
 	RegisterTemp(update)
+	RegisterDeviceList(Temperatures, devs)
 	ts = make(map[string]float32)
 }
 
@@ -17,7 +18,19 @@ func update(temps map[string]int) map[string]error {
 		return map[string]error{"temps": err}
 	}
 	for k, v := range ts {
-		temps[k] = int(v + 0.5)
+		if _, ok := temps[k]; ok {
+			temps[k] = int(v + 0.5)
+		}
 	}
 	return nil
+}
+
+// TODO: Set reasonable default devices
+// CPU (TC[01]P), GPU (TG0P), Memory (Ts0S) and Disk (TH0P)
+func devs() []string {
+	rv := make([]string, len(smc.AppleTemp))
+	for i, v := range smc.AppleTemp {
+		rv[i] = v.Desc
+	}
+	return rv
 }