+++ devices/temp_windows.go
@@ -8,6 +8,7 @@ import (
 
 func init() {
 	RegisterTemp(update)
+	RegisterDeviceList(Temperatures, devs)
 }
 
 func update(temps map[string]int) map[string]error {
@@ -16,9 +17,23 @@ func update(temps map[string]int) map[string]error {
 		return map[string]error{"gopsutil": err}
 	}
 	for _, sensor := range sensors {
-		if sensor.Temperature != 0 {
-			temps[sensor.SensorKey] = int(sensor.Temperature)
+		if _, ok := temps[sensor.SensorKey]; ok {
+			temps[sensor.SensorKey] = int(sensor.Temperature + 0.5)
 		}
 	}
 	return nil
+}
+
+func devs() []string {
+	sensors, err := psHost.SensorsTemperatures()
+	if err != nil {
+		return []string{}
+	}
+	rv := make([]string, 0, len(sensors))
+	for _, sensor := range sensors {
+		if sensor.Temperature != 0 {
+			rv = append(rv, sensor.SensorKey)
+		}
+	}
+	return rv
 }